Do not rely on `cat` and `rm` for consuming the control file, EXIT cleanup
A new command, 'epc consume' was added which will take care of reading
and dumping the control script from Python and truncating it to length 0
(but not **removing** it).

In addition, 'epc consume' is also able to delete the contents of the
temporary directories together with an EXIT-trap.
Thanks to @ChristianParpart for the idea.